+++ Saturn/src/Saturn/Serialisation/AssetSerialisers.cpp
@@ -368,7 +368,7 @@ namespace Saturn {
 						PinKind kind = (PinKind) input[ "Kind" ].as<int>();
 						PinType type = StringToPinType( input[ "Type" ].as<std::string>() );
 
-						Pin pin( pinID, name.c_str(), type, pNewNode->ID );
+						Pin pin( pinID, name.c_str(), type, (int)(size_t)pNewNode->ID );
 						pin.Kind = kind;
 						pin.Node = pNewNode;
