+++ Saturn/src/Saturn/Vulkan/Framebuffer.cpp
@@ -207,7 +207,7 @@ namespace Saturn {
 		VkFramebufferCreateInfo FramebufferCreateInfo = { VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO };
 		FramebufferCreateInfo.renderPass = m_Specification.RenderPass->GetVulkanPass();
 	
-		FramebufferCreateInfo.attachmentCount = m_AttachmentImageViews.size();
+		FramebufferCreateInfo.attachmentCount = (uint32_t)m_AttachmentImageViews.size();
 		FramebufferCreateInfo.pAttachments = m_AttachmentImageViews.data();
 
 		FramebufferCreateInfo.width = m_Specification.Width;
